refactor(cfd):modified the unreasonable imports in the CFD module
