+++ fealpy/cfd/model/stationary_incompressible_navier_stokes_2d.py
@@ -1,13 +1,14 @@
+from typing import Union, Callable, Dict
+
+import sympy as sp
+
 from fealpy.backend import backend_manager as bm
 from fealpy.decorator import cartesian,variantmethod
-from typing import Union, Callable, Dict
-from fealpy.mesh import TriangleMesh
-import sympy as sp
-from fealpy.model.mesher.box_mesher import BoxMesher2d 
 
+from fealpy.mesh import TriangleMesh
+from fealpy.mesher.box_mesher import BoxMesher2d 
 
 CoefType = Union[int, float, Callable]
-
 
 class StationaryNSLFEMPolynomialPDE:
     def __init__(self):
