+++ fealpy/cfd/model/stationary_incompressible_navier_stokes_3d.py
@@ -1,9 +1,11 @@
+from typing import Union, Callable, Dict
+
+import sympy as sp
+
 from fealpy.backend import backend_manager as bm
 from fealpy.decorator import cartesian,variantmethod
-from typing import Union, Callable, Dict
-import sympy as sp
-from fealpy.model.mesher.box_mesher import BoxMesher3d 
 
+from fealpy.mesher.box_mesher import BoxMesher3d 
 
 CoefType = Union[int, float, Callable]
