+++ fealpy/cfd/problem/Kolmogorov_fuild/siny_exp_data_2d.py
@@ -1,6 +1,7 @@
 from typing import Sequence
-from ...backend import backend_manager as bm
-from ...backend import TensorLike
+
+from fealpy.backend import backend_manager as bm
+from fealpy.backend import TensorLike
 
 class SinyExpData2D:
     """
