+++ fealpy/cfd/problem/abc_fuild/sin_cos_data_3d.py
@@ -1,6 +1,7 @@
 from typing import Sequence
-from ...backend import backend_manager as bm
-from ...backend import TensorLike
+
+from fealpy.backend import backend_manager as bm
+from fealpy.backend import TensorLike
 
 class SinCosData3D:
     """
