+++ fealpy/cfd/simulation/fem/cahn_hilliard.py
@@ -1,11 +1,10 @@
 from fealpy.backend import backend_manager as bm
-from ....fem import (ScalarMassIntegrator, ScalarConvectionIntegrator,PressWorkIntegrator,
+from fealpy.decorator import barycentric
+
+from fealpy.fem import (ScalarMassIntegrator, ScalarConvectionIntegrator,PressWorkIntegrator,
                      ViscousWorkIntegrator, SourceIntegrator, GradSourceIntegrator, 
                      ScalarDiffusionIntegrator)
-from ....fem import (BilinearForm, LinearForm, BlockForm, LinearBlockForm)
-from ....decorator import barycentric
-
-
+from fealpy.fem import (BilinearForm, LinearForm, BlockForm, LinearBlockForm)
 
 class CahnHilliardModel:
     def __init__(self, equation, space):
