+++ fealpy/cfd/simulation/fem/stationary_incompressible_ns/stokes.py
@@ -1,11 +1,14 @@
-from .....backend import backend_manager as bm
-from .....fem import LinearForm, BilinearForm, BlockForm, LinearBlockForm
-from .....fem import DirichletBC
-from .....fem import (ScalarMassIntegrator,
+from fealpy.backend import backend_manager as bm
+from fealpy.decorator import barycentric
+
+from fealpy.fem import LinearForm, BilinearForm, BlockForm, LinearBlockForm
+from fealpy.fem import DirichletBC
+from fealpy.fem import (ScalarMassIntegrator,
                      PressWorkIntegrator, ScalarDiffusionIntegrator,
                      SourceIntegrator)
-from .....decorator import barycentric
-from .....solver import spsolve
+
+from fealpy.solver import spsolve
+
 from ..fem_base import FEM
 from ...simulation_base import SimulationBase, SimulationParameters
